refactor: :art: Refactored and seperated modal
