+++ packages/core/src/App/Containers/Modals/app-modals.jsx
@@ -9,6 +9,7 @@ import TradingAssessmentExistingUser from './trading-assessment-existing-user.js
 import CompletedAssessmentModal from './completed-assessment-modal.jsx';
 import DerivRealAccountRequiredModal from 'App/Components/Elements/Modals/deriv-real-account-required-modal.jsx';
 import ExitTradersHubModal from './exit-traders-hub-modal';
+import RiskAcceptTestWarningModal from './risk-accept-test-warning-modal';
 
 const AccountSignupModal = React.lazy(() =>
     moduleLoader(() => import(/* webpackChunkName: "account-signup-modal" */ '../AccountSignupModal'))
@@ -65,12 +66,13 @@ const AppModals = ({
     is_trading_assessment_for_new_user_enabled,
     fetchFinancialAssessment,
     setCFDScore,
-    cfd_score,
     content_flag,
     active_account_landing_company,
     is_deriv_account_needed_modal_visible,
     is_warning_scam_message_modal_visible,
     is_exit_traders_hub_modal_visible,
+    is_trading_experience_incomplete,
+    should_show_risk_accept_modal,
 }) => {
     const url_params = new URLSearchParams(useLocation().search);
     const url_action_param = url_params.get('action');
@@ -114,7 +116,7 @@ const AppModals = ({
         is_logged_in &&
         active_account_landing_company === 'maltainvest' &&
         !is_trading_assessment_for_new_user_enabled &&
-        cfd_score === 0 &&
+        is_trading_experience_incomplete &&
         content_flag !== ContentFlag.LOW_RISK_CR_EU &&
         content_flag !== ContentFlag.LOW_RISK_CR_NON_EU
     ) {
@@ -141,6 +143,8 @@ const AppModals = ({
         ComponentToLoad = <DerivRealAccountRequiredModal />;
     } else if (is_exit_traders_hub_modal_visible) {
         ComponentToLoad = <ExitTradersHubModal />;
+    } else if (should_show_risk_accept_modal) {
+        ComponentToLoad = <RiskAcceptTestWarningModal />;
     }
 
     return (
@@ -164,7 +168,6 @@ export default connect(({ client, ui, traders_hub }) => ({
     has_maltainvest_account: client.has_maltainvest_account,
     fetchFinancialAssessment: client.fetchFinancialAssessment,
     setCFDScore: client.setCFDScore,
-    cfd_score: client.cfd_score,
     setShouldShowVerifiedAccount: ui.setShouldShowVerifiedAccount,
     should_show_cooldown_modal: ui.should_show_cooldown_modal,
     should_show_assessment_complete_modal: ui.should_show_assessment_complete_modal,
@@ -174,4 +177,6 @@ export default connect(({ client, ui, traders_hub }) => ({
     is_warning_scam_message_modal_visible: ui.is_warning_scam_message_modal_visible,
     is_exit_traders_hub_modal_visible: ui.is_exit_traders_hub_modal_visible,
     content_flag: traders_hub.content_flag,
+    is_trading_experience_incomplete: client.is_trading_experience_incomplete,
+    should_show_risk_accept_modal: ui.should_show_risk_accept_modal,
 }))(AppModals);
